Project restructuring for separation of concerns: created src folder for main script and csv files, created utils folder for scraping functions
